+++ apps/core/internal/supervisor/db/gen/const.go
@@ -1,0 +1,6 @@
+package superdb
+
+const (
+	NodeKindFolder = "FOLDER"
+	NodeKindScript = "SCRIPT"
+)